+++ src/Mvc/Mvc.NewtonsoftJson/src/NewtonsoftJsonInputFormatter.cs
@@ -226,8 +226,8 @@ namespace Microsoft.AspNetCore.Mvc.Formatters
                         }
                         else
                         {
-                            addMember = !path.EndsWith($".{member}", StringComparison.Ordinal)
-                                && !path.EndsWith($"[{member}]", StringComparison.Ordinal);
+                            addMember = !path.EndsWith("." + member, StringComparison.Ordinal)
+                                && !path.EndsWith("[" + member + "]", StringComparison.Ordinal);
                         }
                     }
                 }
@@ -242,7 +242,7 @@ namespace Microsoft.AspNetCore.Mvc.Formatters
 
                 exception = eventArgs.ErrorContext.Error;
 
-                var metadata = GetPathMetadata(context.Metadata, path);
+                var metadata = ModelNames.GetPathMetadata(context.Metadata, path);
                 var modelStateException = WrapExceptionForModelState(exception);
                 context.ModelState.TryAddModelError(key, modelStateException, metadata);
 
@@ -299,52 +299,6 @@ namespace Microsoft.AspNetCore.Mvc.Formatters
         /// </remarks>
         protected virtual void ReleaseJsonSerializer(JsonSerializer serializer)
             => _jsonSerializerPool.Return(serializer);
-
-        private ModelMetadata GetPathMetadata(ModelMetadata metadata, string path)
-        {
-            var index = 0;
-            while (index >= 0 && index < path.Length)
-            {
-                if (path[index] == '[')
-                {
-                    // At start of "[0]".
-                    if (metadata.ElementMetadata == null)
-                    {
-                        // Odd case but don't throw just because ErrorContext had an odd-looking path.
-                        break;
-                    }
-
-                    metadata = metadata.ElementMetadata;
-                    index = path.IndexOf(']', index);
-                }
-                else if (path[index] == '.' || path[index] == ']')
-                {
-                    // Skip '.' in "prefix.property" or "[0].property" or ']' in "[0]".
-                    index++;
-                }
-                else
-                {
-                    // At start of "property", "property." or "property[0]".
-                    var endIndex = path.IndexOfAny(new[] { '.', '[' }, index);
-                    if (endIndex == -1)
-                    {
-                        endIndex = path.Length;
-                    }
-
-                    var propertyName = path.Substring(index, endIndex - index);
-                    if (metadata.Properties[propertyName] == null)
-                    {
-                        // Odd case but don't throw just because ErrorContext had an odd-looking path.
-                        break;
-                    }
-
-                    metadata = metadata.Properties[propertyName];
-                    index = endIndex;
-                }
-            }
-
-            return metadata;
-        }
 
         private Exception WrapExceptionForModelState(Exception exception)
         {
